remove service client
