+++ gen/LibKubernetesGenerator/templates/IKubernetes.cs.template
@@ -17,19 +17,8 @@ namespace k8s
 
     /// <summary>
     /// </summary>
-    public partial interface IKubernetes : System.IDisposable
+    public partial interface IKubernetes
     {
-        /// <summary>
-        /// The base URI of the service.
-        /// </summary>
-        System.Uri BaseUri { get; set; }
-
-        /// <summary>
-        /// Subscription credentials which uniquely identify client
-        /// subscription.
-        /// </summary>
-        ServiceClientCredentials Credentials { get; }
-
         {{#.}}
         /// <summary>
         /// {{ToXmlDoc operation.description}}
