+++ src/KubernetesClient/Autorest/HttpExtensions.cs
@@ -10,29 +10,8 @@ namespace k8s.Autorest
     /// <summary>
     /// Extensions for manipulating HTTP request and response objects.
     /// </summary>
-    public static class HttpExtensions
+    internal static class HttpExtensions
     {
-        /// <summary>
-        /// Formats an HttpContent object as String.
-        /// </summary>
-        /// <param name="content">The HttpContent to format.</param>
-        /// <returns>The formatted string.</returns>
-        public static string AsString(this HttpContent content)
-        {
-            if (content != null)
-            {
-                // Await for the content.
-                return
-                    content
-                        .ReadAsStringAsync()
-                        .ConfigureAwait(false)
-                        .GetAwaiter()
-                        .GetResult();
-            }
-
-            return null;
-        }
-
         /// <summary>
         /// Get the content headers of an HtttRequestMessage.
         /// </summary>
@@ -61,80 +40,6 @@ namespace k8s.Autorest
             }
 
             return null;
-        }
-
-        /// <summary>
-        /// Returns string representation of a HttpRequestMessage.
-        /// </summary>
-        /// <param name="httpRequest">Request object to format.</param>
-        /// <returns>The string, formatted into curly braces.</returns>
-        public static string AsFormattedString(this HttpRequestMessage httpRequest)
-        {
-            if (httpRequest == null)
-            {
-                throw new ArgumentNullException("httpRequest");
-            }
-
-            StringBuilder stringBuilder = new StringBuilder();
-            stringBuilder.AppendLine(httpRequest.ToString());
-            if (httpRequest.Content != null)
-            {
-                stringBuilder.AppendLine();
-                stringBuilder.AppendLine("Body:");
-                stringBuilder.AppendLine("{");
-                stringBuilder.AppendLine(httpRequest.Content.ReadAsStringAsync().ConfigureAwait(false).GetAwaiter().GetResult());
-                stringBuilder.AppendLine("}");
-            }
-
-            return stringBuilder.ToString();
-        }
-
-        /// <summary>
-        /// Returns string representation of a HttpResponseMessage.
-        /// </summary>
-        /// <param name="httpResponse">Response object to format.</param>
-        /// <returns>The string, formatted into curly braces.</returns>
-        public static string AsFormattedString(this HttpResponseMessage httpResponse)
-        {
-            if (httpResponse == null)
-            {
-                throw new ArgumentNullException("httpResponse");
-            }
-
-            var stringBuilder = new StringBuilder();
-            stringBuilder.AppendLine(httpResponse.ToString());
-            if (httpResponse.Content != null)
-            {
-                stringBuilder.AppendLine();
-                stringBuilder.AppendLine("Body:");
-                stringBuilder.AppendLine("{");
-                stringBuilder.AppendLine(httpResponse.Content.ReadAsStringAsync().ConfigureAwait(false).GetAwaiter().GetResult());
-                stringBuilder.AppendLine("}");
-            }
-
-            return stringBuilder.ToString();
-        }
-
-        /// <summary>
-        /// Converts given dictionary into a log string.
-        /// </summary>
-        /// <typeparam name="TKey">The dictionary key type.</typeparam>
-        /// <typeparam name="TValue">The dictionary value type.</typeparam>
-        /// <param name="dictionary">The dictionary object.</param>
-        /// <returns>The string, formatted into curly braces.</returns>
-        public static string AsFormattedString<TKey, TValue>(this IDictionary<TKey, TValue> dictionary)
-        {
-            if (dictionary == null)
-            {
-                return "{}";
-            }
-
-            return "{" + string.Join(
-                ",",
-                dictionary.Select(kv => kv.Key.ToString() +
-                                        "=" +
-                                        (kv.Value == null ? string.Empty : kv.Value.ToString()))
-                    .ToArray()) + "}";
         }
     }
 }
