--- src/KubernetesClient/Autorest/ServiceClient.cs
@@ -1,135 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Reflection;
-namespace k8s.Autorest
-{
-    /// <summary>
-    /// ServiceClient is the abstraction for accessing REST operations and their payload data types..
-    /// </summary>
-    /// <typeparam name="T">Type of the ServiceClient.</typeparam>
-    public abstract class ServiceClient<T> : IDisposable
-        where T : ServiceClient<T>
-    {
-        /// <summary>
-        /// Indicates whether the ServiceClient has been disposed.
-        /// </summary>
-        private bool _disposed;
-
-        /// <summary>
-        /// Reference to the first HTTP handler (which is the start of send HTTP
-        /// pipeline).
-        /// </summary>
-        protected HttpMessageHandler FirstMessageHandler { get; set; }
-
-        /// <summary>
-        /// Reference to the innermost HTTP handler (which is the end of send HTTP
-        /// pipeline).
-        /// </summary>
-        protected HttpClientHandler HttpClientHandler { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ServiceClient{T}"/> class.
-        /// </summary>
-        protected ServiceClient()
-            : this(CreateRootHandler())
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ServiceClient{T}"/> class.
-        /// </summary>
-        /// <param name="handlers">List of handlers from top to bottom (outer handler is the first in the list)</param>
-        protected ServiceClient(params DelegatingHandler[] handlers)
-            : this(CreateRootHandler(), handlers)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ServiceClient{T}"/> class.
-        /// Initializes ServiceClient using base HttpClientHandler and list of handlers.
-        /// </summary>
-        /// <param name="rootHandler">Base HttpClientHandler.</param>
-        /// <param name="handlers">List of handlers from top to bottom (outer handler is the first in the list)</param>
-        protected ServiceClient(HttpClientHandler rootHandler, params DelegatingHandler[] handlers)
-        {
-        }
-
-        /// <summary>
-        /// Create a new instance of the root handler.
-        /// </summary>
-        /// <returns>HttpClientHandler created.</returns>
-        protected static HttpClientHandler CreateRootHandler()
-        {
-            // Create our root handler
-#if NET45
-            return new WebRequestHandler();
-#else
-            return new HttpClientHandler();
-#endif
-        }
-
-        /// <summary>
-        /// Gets the HttpClient used for making HTTP requests.
-        /// </summary>
-        public HttpClient HttpClient { get; protected set; }
-
-        /// <summary>
-        /// Gets the UserAgent collection which can be augmented with custom
-        /// user agent strings.
-        /// </summary>
-        public virtual HttpHeaderValueCollection<ProductInfoHeaderValue> UserAgent
-        {
-            get { return HttpClient.DefaultRequestHeaders.UserAgent; }
-        }
-
-        /// <summary>
-        /// Get the HTTP pipelines for the given service client.
-        /// </summary>
-        /// <returns>The client's HTTP pipeline.</returns>
-        public virtual IEnumerable<HttpMessageHandler> HttpMessageHandlers
-        {
-            get
-            {
-                var handler = FirstMessageHandler;
-
-                while (handler != null)
-                {
-                    yield return handler;
-
-                    DelegatingHandler delegating = handler as DelegatingHandler;
-                    handler = delegating != null ? delegating.InnerHandler : null;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Dispose the ServiceClient.
-        /// </summary>
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Dispose the HttpClient and Handlers.
-        /// </summary>
-        /// <param name="disposing">True to release both managed and unmanaged resources; false to releases only unmanaged resources.</param>
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!_disposed)
-            {
-                _disposed = true;
-
-                // Dispose the client
-                HttpClient?.Dispose();
-                HttpClient = null;
-                FirstMessageHandler = null;
-                HttpClientHandler = null;
-            }
-        }
-    }
-}
