--- src/KubernetesClient/IKubernetes.HttpClient.cs
@@ -1,12 +1,0 @@
-using System.Net.Http;
-
-namespace k8s
-{
-    public partial interface IKubernetes
-    {
-        /// <summary>
-        /// Gets the <see cref="HttpClient"/> used for making HTTP requests.
-        /// </summary>
-        HttpClient HttpClient { get; }
-    }
-}
