+++ src/KubernetesClient/IKubernetes.cs
@@ -1,0 +1,15 @@
+namespace k8s;
+
+public partial interface IKubernetes: IDisposable
+{
+    /// <summary>
+    /// The base URI of the service.
+    /// </summary>
+    Uri BaseUri { get; set; }
+
+    /// <summary>
+    /// Subscription credentials which uniquely identify client
+    /// subscription.
+    /// </summary>
+    //ServiceClientCredentials Credentials { get; }
+}
